TMAIL-203: IP warm-up admin + RETURNING-type fix
Final admin manager (closes the TMAIL-196 story). Three sections:

- Tracked IPs table with day/week, today/limit progress bar, total sent,
  start date, state badge.
- Inline 'Start tracking' form posting an IP address to /admin/warmup/start.
- Static 8-week schedule pulled from /admin/warmup/schedule with the
  per-week daily limit and description text the backend already maintains.

Pre-existing backend bug fixed inline: start_warmup() bound the
RETURNING 1 result as i64, but Postgres types '1' as INT4, so every call
500'd with a sqlx type-mismatch and warm-up tracking could never start.
Switch to i32.

Trace baseline drops 33 → 30 — closes the TMAIL-196 admin UI story
entirely (every /api/admin/* route the audit flagged as missing a
consumer now has one).

Verified end to end:
- 1118/1118 vitest still green.
- e2e/specs/admin-warmup-flow.spec.ts: signs up + promotes admin, opens
  /admin/warmup, asserts the 8-week schedule renders, submits
  198.51.100.x via the form, asserts POST 201 + the new IP appears in the
  tracked-IPs table.

+++ backend/src/handlers/warmup.rs
@@ -89,8 +89,10 @@ pub async fn start_warmup(
         )));
     }
 
-    // Added: Insert warm-up tracking record (conflict = already tracking)
-    let result = sqlx::query_scalar::<_, i64>(
+    // Fix: TMAIL-203 — `RETURNING 1` types as INT4 in Postgres. Original
+    // i64 binding 500'd every call with a sqlx type-mismatch. Match the
+    // actual return type.
+    let result = sqlx::query_scalar::<_, i32>(
         "INSERT INTO ip_warmup_tracking (ip_address, current_day, emails_sent_today, total_emails_sent) \
          VALUES ($1, 1, 0, 0) \
          ON CONFLICT (ip_address) DO NOTHING \
